server: split diskspace into platform files to avoid lint conflicts
Bsize is int32 on darwin and int64 on linux. Using int64() cast
triggers unconvert on linux, and nolint:unconvert triggers nolintlint
on macOS. Split into separate files to use the correct type on each.

+++ server/diskspace_darwin.go
@@ -1,5 +1,3 @@
-//go:build !windows && !darwin
-
 package server
 
 import "syscall"
@@ -11,5 +9,5 @@ func availableBytes(path string) (int64, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0, err
 	}
-	return int64(stat.Bavail) * stat.Bsize, nil
+	return int64(stat.Bavail) * int64(stat.Bsize), nil
 }
